new file:   flipper_ml/.gitignore 	modified:   flipper_ml/application.fam 	deleted:    flipper_ml/flipper_ml.c 	new file:   flipper_ml/flipper_ml.cc 	modified:   flipper_ml/lib/tflm/tensorflow/lite/micro/fz/debug_log.cc 	modified:   flipper_ml/lib/tflm/tensorflow/lite/micro/kernels/lstm_eval_common.cc 	new file:   flipper_ml/model/hello_world_int8.tflite 	new file:   flipper_ml/model/model.cc 	new file:   flipper_ml/model/model.h

+++ flipper_ml/lib/tflm/tensorflow/lite/micro/kernels/lstm_eval_common.cc
@@ -191,8 +191,8 @@ CellStateInfo CreateLstmCellStateInfo(const float cell_state_scale,
   cell_state_info.quantized_cell_clip = static_cast<int16_t>(
       std::min(std::max(static_cast<double>(cell_clip) /
                             static_cast<double>(cell_state_scale),
-                        -32768.0),
-               32767.0));
+                        (double)-32768.0),
+               (double)32767.0));
 
   return cell_state_info;
 }
